Segment analysis API moved to v3
