+++ lib/Routes/api_v3_routes.php
@@ -71,3 +71,6 @@ $klein->with( '/api/v3/payable_rate', function () {
 $klein->with( '/api/v3/tm-keys', function () {
     route( '/list', 'GET', '\API\V3\TmKeyManagementController', 'getByUser' );
 } );
+
+route( '/api/v3/projects/[:id_project]/[:password]/segment-analysis',  'GET',  'API\V3\SegmentAnalysisController', 'project' );
+route( '/api/v3/jobs/[:id_job]/[:password]/segment-analysis',          'GET',  'API\V3\SegmentAnalysisController', 'job' );
